requirements.txt
openai
streamlit
azure-identity
azure-search-documents
python-dotenv